+++ src/App.tsx
@@ -1,33 +1,18 @@
-import React from 'react';
+import React from "react";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
-interface Weather {
-  info: {
-    tempMax: number;
-    tempMin: number;
-    type: 'Cloudy' | 'Clear';
-    isEndOfWorld?: boolean;
-  };
+import { HomePage } from "./pages/Home";
+import { ServicesTestsPage } from "./pages/ServicesTests";
+
+function App() {
+  return (
+    <Router>
+      <div className="App">
+        <Switch>
+          <Route path="/" exact component={HomePage} />
+          <Route path="/test" component={ServicesTestsPage} />
+        </Switch>
+      </div>
+    </Router>
+  );
 }
-
-const Weather: React.FC<Weather> = ({ info }) => {
-  return (
-    <div>
-      <p>{info.type}</p>
-    </div>
-  );
-};
-
-const App: React.FC = () => {
-  return (
-    <Weather
-      info={{
-        tempMax: 39.4,
-        tempMin: 32.6,
-        type: 'Cloudy',
-        isEndOfWorld: true,
-      }}
-    />
-  );
-};
-
-export default App;
