+++ src/pages/ServicesTests.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const ServicesTestsPage: React.FC = () => {
+  return <h1>Services Tests Page</h1>;
+};
